mod donations, added admin flag, added user/donation, added project/close, added project/filter, locality/update, notification by gmail

+++ src/main/java/ar/edu/unq/desapp/grupoa/backenddesappapi/service/project/ProjectServiceImp.java
@@ -2,18 +2,23 @@ package ar.edu.unq.desapp.grupoa.backenddesappapi.service.project;
 
 import ar.edu.unq.desapp.grupoa.backenddesappapi.controllers.project.requestbody.ProjectBodyPost;
 import ar.edu.unq.desapp.grupoa.backenddesappapi.controllers.project.requestbody.ProjectBodyPut;
+import ar.edu.unq.desapp.grupoa.backenddesappapi.controllers.project.requestbody.ProjectFilter;
 import ar.edu.unq.desapp.grupoa.backenddesappapi.controllers.project.responsebody.ProjectResponseBody;
 import ar.edu.unq.desapp.grupoa.backenddesappapi.controllers.project.responsebody.ProjectResponseBodyList;
 import ar.edu.unq.desapp.grupoa.backenddesappapi.dao.locality.LocalityDAO;
 import ar.edu.unq.desapp.grupoa.backenddesappapi.dao.project.ProjectDAO;
+import ar.edu.unq.desapp.grupoa.backenddesappapi.dao.user.UserDAO;
 import ar.edu.unq.desapp.grupoa.backenddesappapi.exception.InvalidException;
 import ar.edu.unq.desapp.grupoa.backenddesappapi.exception.InvalidOrNullFieldException;
 import ar.edu.unq.desapp.grupoa.backenddesappapi.model.proyect.Locality;
 import ar.edu.unq.desapp.grupoa.backenddesappapi.model.proyect.Project;
+import ar.edu.unq.desapp.grupoa.backenddesappapi.model.user.User;
+import ar.edu.unq.desapp.grupoa.backenddesappapi.service.Requester;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.util.StringUtils;
 
+import java.io.IOException;
 import java.time.LocalDate;
 import java.util.List;
 import java.util.stream.Collectors;
@@ -23,13 +28,34 @@ public class ProjectServiceImp implements ProjectService {
 
     private @Autowired ProjectDAO projectDAO;
     private @Autowired LocalityDAO localityDAO;
+    private @Autowired UserDAO userDAO;
+    private Requester requester = new Requester();
 
     @Override
     public List<ProjectResponseBodyList> listAllProjects() {
-        return ((List<Project>) projectDAO.findAll())
-                .stream()
-                .map(ProjectResponseBodyList::new)
-                .collect(Collectors.toList());
+            return ((List<Project>) projectDAO.findAll())
+                    .stream()
+                    .map(ProjectResponseBodyList::new)
+                    .collect(Collectors.toList());
+    }
+
+    @Override
+    public List<ProjectResponseBodyList> filterProjects(ProjectFilter body) {
+        if(!StringUtils.hasText(body.getDate())) {
+            return ((List<Project>) projectDAO.findAll())
+                    .stream()
+                    .filter(project -> project.getName().contains(body.getWord()))
+                    .map(ProjectResponseBodyList::new)
+                    .collect(Collectors.toList());
+        }
+        else{
+            return ((List<Project>) projectDAO.findAll())
+                    .stream()
+                    .filter(project -> project.getDeadline().isAfter(LocalDate.parse(body.getDate())))
+                    .filter(project -> project.getName().contains(body.getWord()))
+                    .map(ProjectResponseBodyList::new)
+                    .collect(Collectors.toList());
+        }
     }
 
     @Override
@@ -41,7 +67,10 @@ public class ProjectServiceImp implements ProjectService {
     }
 
     @Override
-    public Integer save(ProjectBodyPost body) throws InvalidOrNullFieldException, InvalidException {
+    public Integer save(ProjectBodyPost body, Long userId) throws InvalidOrNullFieldException, InvalidException {
+        this.validateAdminId(userId);
+        this.validateIsAdmin(userId);
+
         this.validateNewProjectBody(body);
         Locality locality = localityDAO.findById(body.getLocalityId()).orElse(new Locality());
         Project project = new Project();
@@ -49,18 +78,23 @@ public class ProjectServiceImp implements ProjectService {
     }
 
     @Override
-    public void delete(Integer id) throws InvalidException {
+    public void delete(Integer id, Integer userId) throws InvalidException {
         Long value = Long.valueOf(id);
         validateId(value);
+        Long adminId = Long.valueOf(userId);
+        validateAdminId(adminId);
+        validateIsAdmin(adminId);
         projectDAO.deleteById(value);
     }
 
     @Override
-    public ProjectResponseBody update(ProjectBodyPut project, Long id) throws InvalidException, InvalidOrNullFieldException {
-        Long value = Long.valueOf(id);
-        this.validateId(value);
+    public ProjectResponseBody update(ProjectBodyPut project, Long id, Long userId) throws InvalidException, InvalidOrNullFieldException {
+        this.validateId(id);
 
-        Project recoveredProject = projectDAO.findById(value).orElse(new Project());
+        this.validateAdminId(userId);
+        this.validateIsAdmin(userId);
+
+        Project recoveredProject = projectDAO.findById(id).orElse(new Project());
 
         this.validateName(project.getName());
         this.validateFantasyName(project.getFantasyName());
@@ -72,6 +106,39 @@ public class ProjectServiceImp implements ProjectService {
         projectDAO.save(recoveredProject);
 
         return new ProjectResponseBody(recoveredProject);
+    }
+
+    @Override
+    public ProjectResponseBody closeProject(Long id, Long userId) throws InvalidException, IOException {
+        validateAdminId(userId);
+        validateIsAdmin(userId);
+        validateId(id);
+        Project recoverProject = projectDAO.findById(id).orElse(new Project());
+        recoverProject.closeProject();
+        projectDAO.save(recoverProject);
+        recoverProject.getDonations().forEach(donation -> {
+            try {
+                String subject = "the project "+ recoverProject.getName()+ "has been close";
+                String message = "this mail is to inform you that the project "+ recoverProject.getName() + " will be close";
+                this.requester.requestToNotify(subject, message, donation.getEmail());
+            } catch (IOException e) {
+                e.printStackTrace();
+            }
+        });
+        return new ProjectResponseBody(recoverProject);
+    }
+
+    private void validateAdminId(Long adminId) throws InvalidException {
+        if(!userDAO.existsById(adminId)){
+            throw new InvalidException("id: "+adminId);
+        }
+    }
+
+    private void validateIsAdmin(Long adminId) throws InvalidException {
+        User recoveredUser = userDAO.findById(adminId).orElse(new User());
+        if(!recoveredUser.getIsAdmin()){
+            throw new InvalidException("error: permission denied to id "+adminId);
+        }
     }
 
     private void validateId(Long id) throws InvalidException {
